fix(arr): satisfy psalm + drop composer version field
- Cast curl_getinfo() http code result to int so downstream string-concat
  no longer trips MixedOperand.
- Drop the redundant @var array<mixed,mixed> on $decoded = json_decode()
  so is_array() can actually narrow the type (was DocblockTypeContradiction).
- Mark @var array<string, mixed> at return sites of Bazarr/Prowlarr/Radarr/
  Sonarr client methods that surface JSON objects through the generic
  array<mixed,mixed> get()/post() helpers.
- @psalm-suppress MixedAssignment on the two $id = $response['id'] ?? null
  reads in RadarrClient that intentionally accept mixed (filtered by the
  is_numeric() check immediately after).
- Drop the composer.json "version" field — tags provide the version, and
  keeping it triggers a composer-validate warning on CI.

+++ src/Arr/BazarrClient.php
@@ -84,6 +84,7 @@ class BazarrClient
             'language' => $languageCode,
         ];
 
+        /** @var array<string, mixed> */
         return $this->post('/api/v1/subtitles/download', $payload);
     }
 
@@ -140,7 +141,7 @@ class BazarrClient
 
         /** @var string|false */
         $responseBody = curl_exec($ch);
-        $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
+        $httpCode = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
         $curlErrno = curl_errno($ch);
         $curlError = curl_error($ch);
         curl_close($ch);
@@ -165,7 +166,6 @@ class BazarrClient
             return [];
         }
 
-        /** @var array<mixed, mixed> */
         $decoded = json_decode($responseBody, true);
         if (!is_array($decoded)) {
             throw new RuntimeException('Invalid JSON response from Bazarr');
@@ -208,7 +208,7 @@ class BazarrClient
 
         /** @var string|false */
         $responseBody = curl_exec($ch);
-        $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
+        $httpCode = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
         $curlErrno = curl_errno($ch);
         $curlError = curl_error($ch);
         curl_close($ch);
@@ -233,7 +233,6 @@ class BazarrClient
             return [];
         }
 
-        /** @var array<mixed, mixed> */
         $decoded = json_decode($responseBody, true);
         if (!is_array($decoded)) {
             throw new RuntimeException('Invalid JSON response from Bazarr');
